do changes

+++ src/components/GarageDetails.js
@@ -6,7 +6,7 @@ export default class UserDetails extends Component {
     super(props);
 
     this.state = {
-      garages: []
+      garages: [],
     };
 
   }
@@ -28,18 +28,60 @@ export default class UserDetails extends Component {
   }
 
   onDelete = (id) => {
-    axios.delete(`/garage/delete/${id}`).then((res) => {
-      alert("User Delete Successfully");
-      this.retrievePosts();
-    })
-  }
+    if (window.confirm("Are you sure you want to delete this Garage?")) {
+      axios.delete(`/garage/delete/${id}`).then((res) => {
+        alert("Garage Deleted Successfully");
+        this.retrievePosts();
+      });
+    } else {
+      console.log("Garage delete cancelled.");
+    }
+  };
+
+  onTrue = (id,email) => {
+    const data = {
+      access: "True",
+      email: email
+    };
+    if (window.confirm("Are you sure you want to update the part?")) {
+      axios.put("/garage/updateAT/" + id, data).then((res) => {
+        if (res.data.success) {
+          
+          alert("Part Updated Successfully");
+
+        }
+      });
+    }
+  };
+
+  onFalse = (id,email) => {
+    const data = {
+      access: "False",
+      email: email
+    };
+    if (window.confirm("Are you sure you want to update the part?")) {
+      axios.put("/garage/updateAF/" + id, data).then((res) => {
+        if (res.data.success) {
+          
+          alert("Part Updated Successfully");
+
+        }
+      });
+    }
+  };
+
 
   filterData(garages, searchKey) {
 
     const result = garages.filter((post) =>
       post.name.toLowerCase().includes(searchKey) ||
-      post.email.toLowerCase().includes(searchKey) ||
-      post.phone.toLowerCase().includes(searchKey)
+      post.town.toLowerCase().includes(searchKey) ||
+      post.address.toLowerCase().includes(searchKey) ||
+      post.number.toLowerCase().includes(searchKey) ||
+      post.latitude.toLowerCase().includes(searchKey) ||
+      post.longitude.toLowerCase().includes(searchKey) ||
+      post.access.toLowerCase().includes(searchKey) ||
+      post.email.toLowerCase().includes(searchKey)
     )
 
     this.setState({ garages: result })
@@ -74,7 +116,7 @@ export default class UserDetails extends Component {
         </div>
 
         <div className="row">
-          <div className="col-lg-3 mt-2 mb-2">
+          <div className="col-lg-12 mt-12 mb-12">
             <input
               className="form-control"
               type="search"
@@ -85,7 +127,8 @@ export default class UserDetails extends Component {
           </div>
         </div>
 
-        <table class="table" style={{ marginBottom: "60px" }}>
+        <table class="table" style={{ marginBottom: "60px", marginTop: "20px" }}>
+
           <thead>
             <tr>
               <th scope="col"> </th>
@@ -97,6 +140,7 @@ export default class UserDetails extends Component {
               <th scope="col">Longitude</th>
               <th scope="col">Access</th>
               <th scope="col">Owner E-Mail</th>
+              <th scope="col">Garage Photo</th>
             </tr>
           </thead>
           <tbody>
@@ -116,13 +160,13 @@ export default class UserDetails extends Component {
                 <td>{garages.access}</td>
                 <td>{garages.email}</td>
                 <td>
-                  <img alt="" className="activator" style={{ width: 200, height: 200 }} src={'data:image/jpg;base64,' + garages.image1} />
+                  <img alt="profile" className="activator" style={{ width: 150, height: 150 }} src={'data:image/jpg;base64,' + garages.image1} />
                 </td>
                 <td className="text-center">
-                  <div className="d-flex text-center">
-                    <button className="btn btn-primary" href={`/garage/edit/${garages._id}`}>
-                      <i className="fas fa-edit"></i>&nbsp;Edit All
-                    </button>
+                  <div className="text-center">
+                  <a className="btn btn-primary" href={`/garage/edit/${garages._id}`}>
+                    <i className="fas fa-edit"></i>&nbsp;Edit All
+                  </a>
                     &nbsp;
                     &nbsp;
                     <button className="btn btn-danger" onClick={() => this.onDelete(garages._id)}>
@@ -132,9 +176,16 @@ export default class UserDetails extends Component {
                   &nbsp;
                   &nbsp;
                   <div>
-                    <button className="btn btn-primary" href={`/garage/edit/${garages._id}`}>
-                      <i className="fas fa-edit"></i>&nbsp;Update Access
+                  <div className="text-center">
+                    <button className="btn btn-success" onClick={() => this.onTrue(garages._id,garages.email)}>
+                      <i className="fas fa-edit"></i>&nbsp;Give Access
                     </button>
+                    &nbsp;
+                    &nbsp;
+                    <button className="btn btn-danger" onClick={() => this.onFalse(garages._id,garages.email)}>
+                      <i className="fa fa-trash"></i>&nbsp;Access Denied
+                    </button>
+                  </div>
                   </div>
                 </td>
               </tr>
